# Do not check trace for diffusers, saving memory and time for FLUX