--- a/optimum/exporters/openvino/utils.py
+++ b/optimum/exporters/openvino/utils.py
@@ -15,10 +15,10 @@
 import inspect
 import logging
 from collections import namedtuple
-from pathlib import Path
-from typing import Any, Callable, Dict, List, Optional, Tuple, Union
 from contextlib import contextmanager
 from functools import partial
+from pathlib import Path
+from typing import Any, Callable, Dict, List, Optional, Tuple, Union
 
 from transformers import PretrainedConfig
 from transformers.utils import is_torch_available