# What does this PR do?

This is further optimization of memory consumption of diffusers conversion, continuation of this PR: #1033 
When `check_trace=True` the TorchScript graph is generated second time and is compared with the graph generated first time. It is useful to catch incorrect traced graph sometimes, but in optimum we control which models are supported and such issues shouldn't happen.
Currently only introduce this for `diffusers`, but can be done for all the models.
The most impact on memory is demonstrated for FLUX, for other diffusers it significantly reduces conversion time.

| Model | Before | After |
| ------- | ------- | ------ |
| FLUX | ~58Gb/~240s ![image](https://github.com/user-attachments/assets/2bd9d2a9-6117-4b73-917f-345ed29ac966) | ~33Gb/~130s ![image](https://github.com/user-attachments/assets/87bd5823-5f07-4ee2-8482-18b9803274e3) |
| SD-3.5-medium | ~18Gb/~250s ![image](https://github.com/user-attachments/assets/98f9b416-351f-4e7f-be52-4341dbbb368d) | ~18Gb/~100s ![image](https://github.com/user-attachments/assets/57a6cfc4-614b-41c1-ae67-74f1a9505429) 
 |

Fixes # (issue)


## Before submitting
- [ ] This PR fixes a typo or improves the docs (you can dismiss the other checks if that's the case).
- [ ] Did you make sure to update the documentation with your changes?
- [ ] Did you write any new necessary tests?