@AlexKoff88 2024-12-12
Awesome! Thanks @mvafin.

Please check and fix the quality: 
```
black --check .
ruff check .
```