@HuggingFaceDocBuilderDev 2024-12-12
The docs for this PR live [here](https://moon-ci-docs.huggingface.co/docs/optimum-intel/pr_1064). All of your documentation changes will be reflected on that endpoint. The docs are available until 30 days after the last update.