@mvafin 2024-12-12
> ruff check .

Done