@AlexKoff88 2024-12-12
+ @alexsu52 . Shall we have an option that disables this behavior during model export? Such an option can be enabled by default.