@AlexKoff88 2024-12-13
> Not a fan of patches, long term solution would be decouple the export process by doing `torch.jit.trace` with the arguments we want then `ov.convert_model` with the resulting model.

I agree. I see that torch.jit.trace in one place in the Optimum-Intel code. @mvafin, why did you just not pass the argument there?