@IlyasMoutawwakil 2024-12-17
okay, we can go with a patch but let's at least make it a clean one, the naming in the PR is a bit confusing and I don't think the patching should be in `__main__.py` but rather in `convert` where lower level logic is defined. What about this:
- add an argument `check_trace=None` to the export function(s) so that users can control this behavior.
- if None, set this value to False in the case of diffusers.
- do the patching around the `convert_model` function with `patch_torch_jit_trace(check_trace=check_trace)`.